refactor(types.ts): moving breakpointsProps type to the types file
